BF: Fix ReadmeFrame not showing again after close
After closing (and not just hiding) the ReadmeFrame in Builder it was
not possible to show it again using the "toggle" option in the "view"
menu. The ReadmeFrame removes its reference from parent when closing.
Using the menu entry then caused an AttributeError. This was fixed by
checking the attribute (which exists because of __init__) and calling
updateReadme() and showReadme() if it does not contain a ReadmeFrame.